fix: replace types to fixed size

+++ M1/MT19937.h
@@ -1,19 +1,20 @@
 #ifndef MT19937_H
 #define MT19937_H
 
+#include <stdint.h>
 #include <stddef.h>
 
 #define MT_N 624
 
 typedef struct
 {
-    unsigned long mt[MT_N];
-    size_t        index;
+    uint32_t mt[MT_N];
+    size_t   index;
 }
 MT19937;
 
-void          MT19937_Init(MT19937* gen, unsigned long seed);
-unsigned long MT19937_Rand(MT19937* gen);
-double        MT19937_RandRange(MT19937* gen, double a, double b);
+void     MT19937_Init(MT19937* gen, uint32_t seed);
+uint32_t MT19937_Rand(MT19937* gen);
+double   MT19937_RandRange(MT19937* gen, double a, double b);
 
 #endif
