+++ M1/MonteCarlo.h
@@ -1,9 +1,11 @@
 #ifndef MONTECARLO_H
 #define MONTECARLO_H
 
+#include <stdint.h>
+
 #define mcResult(T) struct { T elapsed_time, area; }
 typedef mcResult(double) mcResultd;
 
-mcResultd MonteCarlo_Calculate(unsigned long radius, unsigned long threads_count, unsigned long tries, unsigned long seed);
+mcResultd MonteCarlo_Calculate(uint32_t radius, uint32_t threads_count, uint32_t tries, uint32_t seed);
 
 #endif
